Move ScanValidator to data_conversion_layer
