+++ standalone/include/psen_scan_v2_standalone/protocol_layer/scanner_state_machine.h
@@ -49,7 +49,7 @@
 #include "psen_scan_v2_standalone/data_conversion_layer/scanner_reply_serialization_deserialization.h"
 #include "psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_msg.h"
 #include "psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_deserialization.h"
-#include "psen_scan_v2_standalone/complete_scan_validator.h"
+#include "psen_scan_v2_standalone/data_conversion_layer/complete_scan_validator.h"
 #include "psen_scan_v2_standalone/util/watchdog.h"
 
 namespace psen_scan_v2_standalone
